Merge in NEDS/harmony-access-point from NEDS-49_rebase to NEDS-develop

* commit '99906b9e14c82932382561e0a75f6bdce0d44467':
  Ignore unit tests
  Ignore unit tests
  Rebase to 4.2.4
  - fixed some failing tests - ignored intermittently failing integration tests (issues with spring context cleaning?)
  - moved most dependencies to central - fixed references to child modules
  removed wildfly and weblogic modules
  created new pom structure
  EDELIVERY-8438 - Changed version to 4.2.4
  [EDELIVERY-8424] fix unit tests
  [EDELIVERY-8424] [EDELIVERY-8237] Fix long int overflow for entity id passed to the UI.  [part of PR #2476]
  [EDELIVERY-8424] change more entityIds from int to long
  [EDELIVERY-8424] mention 4.2.4 in upgrade-info.txt
  [EDELIVERY-8424] change version to 4.2.4-SNAPSHOT
  [EDELIVERY-8424] fix unit test
  [EDELIVERY-8424] forgotten pmode int ids
  [EDELIVERY-8424] [EDELIVERY-8336] put long instead of int for pMode id [PR #2571]
  [EDELIVERY-8424] [EDELIVERY-8239] Fix issue with pMode not showing in the UI [PR #2520]
  [EDELIVERY-8424] [EDELIVERY-8334] fix command propagation on oracle [PR #2553]

+++ Domibus-MSH-ext-services-delegate/src/test/java/eu/domibus/ext/rest/PModeFileExtResourceTest.java
@@ -41,7 +41,7 @@ public class PModeFileExtResourceTest {
 
     @Test
     public void test_downloadPMode(@Mocked ResponseEntity responseEntity) {
-        final int pModeId = 1;
+        final long pModeId = 1;
 
         final byte[] bytes = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n".getBytes(StandardCharsets.UTF_8);
 
@@ -64,7 +64,7 @@ public class PModeFileExtResourceTest {
 
     @Test
     public void test_downloadPMode_NoContent(@Mocked ResponseEntity responseEntity) {
-        final int pModeId = 1;
+        final long pModeId = 1;
 
         final byte[] bytes = "".getBytes(StandardCharsets.UTF_8);
 
@@ -88,7 +88,7 @@ public class PModeFileExtResourceTest {
     @Test
     public void test_GetCurrentPMode(final @Mocked PModeArchiveInfoDTO pModeArchiveInfoDTO) {
 
-        final int pModeId = 2;
+        final long pModeId = 2;
 
         new Expectations() {{
             pModeExtService.getCurrentPmode();
